+++ server/dev-requirements.txt
@@ -1,4 +1,6 @@
 black>=20.8b0
 flake8
 isort
+mkdocs
+mkdocs-material
 pytest
