+++ server/dive_server/pipelines.py
@@ -4,9 +4,9 @@ from typing import Dict, List, Optional
 
 from girder.models.folder import Folder
 
-from viame_server.constants import TrainedPipelineCategory, TrainedPipelineMarker
-from viame_server.utils import get_static_pipelines_path
-from viame_utils.types import PipelineCategory, PipelineDescription
+from dive_server.constants import TrainedPipelineCategory, TrainedPipelineMarker
+from dive_server.utils import get_static_pipelines_path
+from dive_utils.types import PipelineCategory, PipelineDescription
 
 AllowedStaticPipelines = r"^detector_.+|^tracker_.+|^generate_.+"
 DisallowedStaticPipelines = (
