+++ server/dive_server/serializers/meva.py
@@ -11,7 +11,7 @@ from boiler.definitions import ActorType
 from boiler.serialization import kpf
 from girder.models.file import File
 
-from viame_server.serializers.models import Feature, Track
+from dive_server.serializers.models import Feature, Track
 
 
 @dataclass
