+++ server/dive_tasks/__init__.py
@@ -5,11 +5,11 @@ from girder_worker import GirderWorkerPluginABC
 logging.basicConfig(format="%(levelname)s:%(message)s", level=logging.INFO)
 
 
-class ViamePlugin(GirderWorkerPluginABC):
+class DIVEPlugin(GirderWorkerPluginABC):
     def __init__(self, app, *args, **kwargs):
         self.app = app
 
     def task_imports(self):
         # Return a list of python importable paths to the
         # plugin's path directory
-        return ["viame_tasks.tasks"]
+        return ["dive_tasks.tasks"]
