+++ server/tests/test_serialize_viame_csv.py
@@ -5,7 +5,7 @@ from typing import Dict, List
 
 import pytest
 
-from viame_server.serializers import models, viame
+from dive_server.serializers import models, viame
 
 # Test cases can use this by staying under frame 100
 filenames = [f"{str(i)}.png" for i in range(1, 100)]
